+++ app/src/main/java/reikai/presentation/novel/browse/NovelBrowseScreen.kt
@@ -353,7 +353,7 @@ class NovelBrowseScreen(
             is NovelBrowseDialog.ChangeCategory -> NovelCategoryDialog(
                 dialog = NovelDetailsDialog.ChangeCategory(dialog.allCategories, dialog.currentCategoryIds),
                 onDismiss = viewModel::dismissDialog,
-                onConfirm = { viewModel.applyCategories(dialog.novelId, it) },
+                onConfirm = { viewModel.applyCategories(dialog.target, it) },
             )
             is NovelBrowseDialog.RemoveNovel -> EntryRemoveDialog(
                 title = dialog.item.name,